Only fallback to cmd identifier when needed
Using the cmd identifier is quite problematic.  If it contains null
bytes, it will result in an error farther down the line, causing the
group not to be shown!  This change makes cmd act in the same way as the
window title, in that res_class and res_name are preferred.

It is probably wise to split cmd by null bytes and choose only the first
one, but since I am unsure of all the potential ramifications of such a
change (would it be better to replace null bytes with spaces?) and
because this change fixes the problem for me, I will not include that in
this commit.